Adjust Script to produce inputs for Farsite
This has led to changes to:
* Fix gt.nop.rcls (gdal option) - allow zeros as reclassified value; lead better with NoData and DataTypes;
* Improve gdal_mapcalc in gt.nop.alg;
* Add method to convert Float Raster to Integer Raster (gt.nop.alg.floatrst_to_intrst);
* Produce GRASS GIS aspect rasters similars to ArcGIS Results;
* Delete rst_dataType from gt.prop.rst;
* Fix create_fishnet in gt.sample;

Tool for Raster resampling based on majority: RST1 is a raster with cellsize 50 and RST2 is a raster with cellsize 10; we want resample RST2 to make it a raster with cellsize 50, but we want that the value of the new raster be the majority between the values of the 10m cells overlapping one 50m cells.

gt.spnlst (Close #9).